resource service * add storage * int helper method * rename meilisearch to meili * adjust config

+++ internal/data/schema/mixin/int.go
@@ -43,7 +43,7 @@ type Size struct{ mixin.Schema }
 // Fields of the Size mixin.
 func (Size) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("size").
+		field.Int64("size").
 			Default(0).
 			Comment("size in bytes"),
 	}
